ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
anthropic==0.54.0
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
